Refactored profiling module. Added function to write profiling data to JSON. Added method on `Context` to write profiling data to JSON.
