+++ Cargo.toml
@@ -9,7 +9,8 @@ default = ["profiling"]
 profiling = ["humantime"]
 
 [dependencies]
-ixa = "0.3.0"
+#ixa = "0.3.0"
+ixa = { path = "../ixa" }
 csv = "1.3.1"
 tempfile = "3.19.1"
 statrs = "0.18.0"
