+++ input/input.json
@@ -45,6 +45,6 @@
       ],
       "hospital_incidence_report_name": "hospital_incidence_report.csv"
     },
-    "profiling_data_prefix": "profiling_"
+    "profiling_data_path": "profiling_data.json"
   }
 }
